Fix common/test_utils.py on Python 3.5.4+
In CPython commit e59af55c2, instantiating a logging.SysLogHandler
stopped raising an exception if the syslog server was
unavailable. This commit first appears in CPython
3.5.4. utils.get_logger() catches that error and retries the
instantiation, and there a test asserting that. The test fails on
Python 3.5.4 or greater, so now it has been corrected to only assert
things about the first instantiation of logging.SysLogHandler and
passes on Python 3.5.4 and 3.5.5.

This was noticed by running "tox -e py35" on an Ubuntu 18.04 system,
which ships with Python 3.5.5.

Change-Id: I43f231bd7d3566b9849a48f46ec9e2af4cd23be4