Refactor object updater's async-pending finder
Pulled the FS walking out into its own function so that finding
async_pendings is in one place and processing them is in another. This
should make it easier to add greenthread-based concurrency later.

Also refactored rate limiting to use utils.RateLimitedIterator instead
of calling ratelimit_sleep ourselves. The functionality's the same,
but the interface is easier.

Change-Id: I02f9e3201e9434080c6fb87ac5e249d8b327a231