Fix shard_max_row in ContainerBroker.get_replication_info()
ContainerBroker adds the shard_max_row item to the
get_replication_info result by querying the db for the shard ranges
table max rowid. However, the wrong table name was being used in the
db query such that the value was always -1. This bug was benign
because the value of shard_max_row is not currently used.

Noticed while reviewing [1] which does make use of the shard_max_row
*key* in replication info.

[1] Related-Change: I7231e8af310e268484f2075f0194b7783cf1c3ea

Change-Id: I9e733e301894f1ffff4a1092926cc0df8419c5b5