Fix a potential bug
In the class method from_hash_dir(), the arguments to input when creating an instance of the BaseDiskFile class are incorrect.
The __init__() method of BaseDiskFile class receive the arguments in order of mgr, device_path, partition and etc.
However, in from_hash_dir() method, the order of arguments are mgr, device_path, None and partition
The class method from_hash_dir() is used by the Object Auditor.
If the partition argument is used in the new DiskFile implementations, exception may occur.
It will be cause object auditing to failed and the object will be quarantine by the Object Auditor.

Closes-Bug: #1805539
Change-Id: Ic2e29474505426dea77e178bf94d891f150d851b