s3api: fix partition between S3 and Swift headers
In swift3 [1], we partitioned between swift3-specific sysmeta headers and
all other swift headers. During the transition to s3api [2], we were
sure to include a support for reading swift3 sysmeta as s3api sysmeta,
but we forgot to add all *other* sysmeta to the swift headers!

Fix that, and rename sw_sysmeta_headers to s3_sysmeta_headers to
more-accurately reflect its purpose.

[1] https://github.com/openstack/swift3/blob/1.12/swift3/response.py#L90-L96
[2] https://review.openstack.org/#/c/557623/

Change-Id: Id6de002b18ba784ad53bda38855376159c7d2ba7