py3: port object controller in proxy
When looking at the tuple thing, remember that tuples are comparable
with ints in py2, but they do not compare like (100)[0]. Instead, they
are always greater, acting almost like MAX_INT. No wonder py3 banned
that practice.

We had a much of other noncomparables, but we dealt with them in
obvious ways, like adding key= to sorted().

Change-Id: I52e96406c3c1f39b98c1d81bdc057805cd1a6278