encryption: Be more tolerant of errors while decrypting listings
Previously, if a bad root encryption secret was loaded, container
listings may always return 500 Internal Error; logs would show this to
be the result of a UnicodeDecodeError while trying to decrypt an
object's ETag.

When we added support for multiple root encryption secrets, we already
had to tolerate unknown secret IDs, and would replace the "hash" value
with "<unknown>".

Now, do that when the secret ID is known, but the secret is wrong (as
evidenced by the fact that the decrypted hash is not UTF-8).

Change-Id: I17097070af4e395fdb631e30f94db97e1848a577
Related-Change: I40307acf39b6c1cc9921f711a8da55d03924d232
Closes-Bug: 1729921
Co-Authored-By: Alistair Coles <alistairncoles@gmail.com>
Signed-off-by: Tim Burke <tim.burke@gmail.com>