Leave less garbage in /var/tmp
All our tests that invoked broker.set_sharding_state() created
/var/tmp/tmp, when it called DatabaseBroker.get_device_path(),
then added "tmp" to it. We provided 1 less level, so it walked up
ouside of the test's temporary directory.

The case of "cleanUp" instead of "tearDown" didn't break out of
jail, but left trash in /var/tmp all the same.

Change-Id: I8030ea49e2a977ebb7048e1d5dcf17338c1616df