py3:  display help if no subcommand provided to manage-shard-ranges
That way we don't traceback when you just provide a DB file.

At some point, py3 switched to having optional subparsers. In py37,
they added a kwarg to say whether a subparser is optional or required,
but for the sake of earlier versions, we have to check whether it was
set manually and cannot rely on argparse doing it for us. See also:

- https://bugs.python.org/issue9253
- https://github.com/python/cpython/commit/aaf6fc0
- https://bugs.python.org/issue33109
- https://github.com/python/cpython/commit/8ebf5ce

Change-Id: Iac8048ad0e23ffb28b914fe880c4b6538bc41c86