Enable swift-dsvm-functional tests at losf feature branch
In the specific dissimilar branch, jobs that require other projects,
which doesn't have same branch name, will not run in the gerrit gate.
This is the cause why we cannot see swift-dsvm-functional test.

To pull the required project correctly, set override-checkout flag
to get the master branch explicitly. In addition, we also requires
devstack-minimal definition so this patch add devstack repo as a
required-project too.

This commit won't be needed when we are ready to gate the losf branch
into the master.

Change-Id: I7e885e5b4c85ffb1cdbda7c254aade69df57a345