Fix lower-constraint deps handling
When the lower-constraints tox target was added, it was assumed the
install_command was just running the install and that the dependencies
and constraints were being set using "deps = ".

This fixed the install_command and deps to follow the expected pattern
so the lower-constraints job actual does install the lower constraints.

Also:
  * raise the minimum for netifaces as the existing version would fail
    to load
  * pin oslo.log to the minimum required by our minimum
    keystonemiddleware
  * fix up some error handling for ancient versions of pastedeploy
  * specify that we shouldn't install xattr on a platform we don't
    support anyway, because reasons

Change-Id: Ie78c0dabe12e01109db2b6412166c3564b87ee96
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>