Move from py35 to py37 for gating py3 jobs
Looking at https://governance.openstack.org/tc/reference/runtimes/train.html
OpenStack is no longer targetting Python 3.5, so it doesn't really make sense
to gate on it. Move to Python 3.7, just to make it so we can go longer before
having to do this shuffle again.

Presumably, once we actually declare support for Python 3.6, we should
make that job voting, too? We've had enough gate flaky-ness lately that
I'm not keen on it.

Change-Id: I9c9e5df76a6592ab9c512b7c1ba807f9a0053059