Fix ostestr invocation
More-recent versions of stestr are rather picky about wanting an integer
for concurrency [1], but ostestr prior to [2] defaults it to None. As a
result, all our func tests would fail with:

   TypeError: int() argument must be a string or a number, not 'NoneType'

Also, move the rolling upgrade job to non-voting since it's going to
pull in the old version but try to run with new deps.

[1] https://github.com/mtreinish/stestr/commit/c93bc81
[2] https://review.openstack.org/#/c/651275/

Change-Id: Icba6c46bcff0e0516107e0f6742d32c92488ae20