Add losf package installation for gate job
And also calling the setups as pre-run scripts.
To make it working well, this includes changes below:

Change the Makefile to work on vagrant swift all in one

With the $PATH properly set (needs $HOME/go/bin, and the path to cmake,
if needed), it works and compiles swift-rpc-losf.

Also, -Wl,-rpath does not appear to work yet, it would be nice for the
binary not to require setting LD_LIBRARY_PATH to find leveldb.so
It seems that the "go get levigo.." line compiles it but does not
install the binary.

Co-Authored-By: Alexandre Lécuyer <alexandre.lecuyer@corp.ovh.com>

Change-Id: Ief3287e650c80dea951d2c85b16d1eaa9b7631d5