Require logger passed to broker in unit tests for account-reaper
This patch removes default value for logger passed to
faked BrokerClass in some unit tests for account reaper,
to make sure that a logger instance is properly passed
when generation BrokerClass instance.

This patch also removes unnecessory parameter, which is
added to confirm logger passed to broker, but is not
useful in fact.

NOTE: This patch is a follow-up of my previous commit[1]

[1] https://review.opendev.org/#/c/295875/

Change-Id: Ica10e3ac42375c2dc141478e647408df90028ae9