ring: Track more properties of the ring
Plumb the version from the ringbuilder through to the metadata at the
start of the ring. Recover this (if available) when running

    swift-ring-builder <ring> write_builder

When we load the ring, track the count and MD5 of the bytes off disk, as
well as the number of uncompressed bytes.

Expose all this new information as properties on the Ring, along with

  - device_count (number of non-None entries in self._devs),
  - weighted_device_count (number of devices that have weight), and
  - assigned_device_count (number of devices that actually have
    partition assignments).

Co-Authored-By: Matthew Oliver <matt@oliver.net.au>
Change-Id: I73deaf6f1d9c1d37630c37c02c597b8812592351